pyproject.toml
[project]
name = "codebase-reader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.1",
    "fastapi>=0.125.0",
    "gitpython>=3.1.45",
    "google-genai>=1.56.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-text-splitters>=1.1.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "tiktoken>=0.12.0",
    "uvicorn>=0.38.0",
]
